docs: move more boilerplate into DocPage
